+++ SwiftLinkPreviewTests/Constants.swift
@@ -1,0 +1,44 @@
+//
+//  Constants.swift
+//  SwiftLinkPreview
+//
+//  Created by Leonardo Cardoso on 05/07/2016.
+//  Copyright © 2016 leocardz.com. All rights reserved.
+//
+
+import Foundation
+
+struct Constants {
+    
+    static let bodyTextSpan = "body-text-span"
+    static let bodyTextP = "body-text-p"
+    static let bodyTextDiv = "body-text-div"
+    static let bodyTextDivSpanP = "body-text-div-span-p"
+    static let bodyImageSingle = "body-image-single"
+    static let bodyImageGallery = "body-image-gallery"
+    static let headMetaTwitter = "head-meta-twitter"
+    static let headMetaMeta = "head-meta-meta"
+    static let headMetaFacebook = "head-meta-facebook"
+    
+    
+    static let headRandomPre = "[:head-random-pre]"
+    static let headRandomPos = "[:head-random-pos]"
+    
+    static let bodyRandom = "[:body-random]"
+    
+    static let twitterTitle = "[:twitter-title]"
+    static let twitterSite = "[:twitter-site]"
+    static let twitterImageSrc = "[:twitter-image-src]"
+    static let twitterDescription = "[:twitter-description]"
+    
+    static let facebookTitle = "[:og-title]"
+    static let facebookSite = "[:og-url]"
+    static let facebookImage = "[:og-image]"
+    static let facebookDescription = "[:og-description]"
+    
+    static let title = "[:title]"
+    static let site = "[:site]"
+    static let image = "[:image]"
+    static let description = "[:description]"
+    
+}