+++ SwiftLinkPreviewTests/StringExtension.swift
@@ -85,4 +85,20 @@ extension String {
         
     }
     
+    // Replace
+    func replace(search: String, with: String) -> String {
+        
+        if let clearWith: String = with.stringByReplacingOccurrencesOfString(" ", withString: ""),
+            let replaced: String = self.stringByReplacingOccurrencesOfString(search, withString: clearWith) {
+            
+            return replaced
+            
+        } else {
+            
+            return self
+            
+        }
+        
+    }
+    
 }